DockerApi.Container (docker_api v0.5.1)

Functions

all()

all(host)

Fetch all the containers from a given docker host

host: "127.0.0.1"

iex> DockerApi.Container.all("127.0.0.1")

  [%{...}, ..]

all(host, opts)

Fetch all the containers from a given docker host

host: "127.0.0.1" opts: %{}

See docker API documentation for full list of query parameters

iex> DockerApi.Container.all("127.0.0.1", %{all: 1})

[%{...}, ..]

attach(host, id)

changes(host, id)

create(host, opts)

delete(host, id)

delete(host, id, opts)

Delete a container

host: "127.0.0.1" id: "123456" otps: %{}

exec(host, id, opts)

exec_start(host, id, opts)

find(container_id)

find(hosts, id)

Find a container when hosts is a String

hosts: "127.0.0.1" id: "1234567"

iex> DockerApi.Container.find("127.0.0.1", "123456")
  %{...}

kill(host, id)

logs(host, id)

Fetch the logs from a container

Returns the last 50 entries for stdout and stderr.

restart(host, id)

start(host, id)

start(host, id, opts)

stop(host, id)

top(host, id)

Top running processes inside the container